Move hashing utilities into their own crate
I'm about to create a circular dependency which I need to break.
