+++ src/rust/engine/fs/fs_util/Cargo.toml
@@ -9,4 +9,5 @@ boxfuture = { path = "../../boxfuture" }
 clap = "2"
 fs = { path = ".." }
 futures = "0.1.16"
+hashing = { path = "../../hashing" }
 protobuf = "1.4.1"
