+++ src/rust/engine/src/nodes.rs
@@ -15,6 +15,7 @@ use context::Context;
 use core::{Failure, Key, Noop, TypeConstraint, Value, Variants, throw};
 use externs;
 use fs::{self, Dir, File, FileContent, Link, PathGlobs, PathStat, VFS};
+use hashing;
 use rule_graph;
 use selectors::{self, Selector};
 use tasks;
@@ -759,9 +760,9 @@ impl From<ReadLink> for NodeKey {
 pub struct DigestFile(File);
 
 impl Node for DigestFile {
-  type Output = fs::Digest;
+  type Output = hashing::Digest;
 
-  fn run(self, context: Context) -> NodeFuture<fs::Digest> {
+  fn run(self, context: Context) -> NodeFuture<hashing::Digest> {
     let file = self.0.clone();
     context
       .core
@@ -1162,7 +1163,7 @@ impl Node for NodeKey {
 #[derive(Clone, Debug)]
 pub enum NodeResult {
   Unit,
-  Digest(fs::Digest),
+  Digest(hashing::Digest),
   DirectoryListing(DirectoryListing),
   LinkDest(LinkDest),
   Snapshot(fs::Snapshot),
@@ -1187,8 +1188,8 @@ impl From<fs::Snapshot> for NodeResult {
   }
 }
 
-impl From<fs::Digest> for NodeResult {
-  fn from(v: fs::Digest) -> Self {
+impl From<hashing::Digest> for NodeResult {
+  fn from(v: hashing::Digest) -> Self {
     NodeResult::Digest(v)
   }
 }
@@ -1269,7 +1270,7 @@ impl TryFrom<NodeResult> for fs::Snapshot {
   }
 }
 
-impl TryFrom<NodeResult> for fs::Digest {
+impl TryFrom<NodeResult> for hashing::Digest {
   type Err = ();
 
   fn try_from(nr: NodeResult) -> Result<Self, ()> {
